requirements.txt
paho-mqtt~=1.5.0
numpy~=1.19.0
ta-lib
